Refacto CompilationStyle
