+++ src/views/components/compilation/CompilationStyle.tsx
@@ -1,6 +1,7 @@
 import { MessageBoxContainer } from '@components/MessageBoxContainer';
 import { PrimaryButton } from '@components/buttons';
-import { Input } from '@components/inputs';
+import { Input, LoggerInput } from '@components/inputs';
+import { BlockProgressBar } from '@components/progress-bar/ProgressBar';
 import styled from 'styled-components';
 
 export const CompilationDialogContainer = styled(MessageBoxContainer)`
@@ -33,6 +34,7 @@ export const CompilationDialogContainer = styled(MessageBoxContainer)`
     border-top: 1px solid ${({ theme }) => theme.colors.dark20};
     padding-top: 16px;
     align-items: center;
+    justify-content: space-between;
 
     .executable-info {
       ${({ theme }) => theme.fonts.normalRegular}
@@ -114,5 +116,106 @@ export const CompilationOptionsContainer = styled.div`
       display: flex;
       align-items: center;
     }
+  }
+`;
+
+export const CompilationLogsContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+  user-select: none;
+  height: 100%;
+  width: 100%;
+
+  .title {
+    ${({ theme }) => theme.fonts.titlesHeadline6};
+    ${({ theme }) => theme.colors.text100};
+  }
+
+  .logs {
+    display: flex;
+    flex-direction: column;
+    gap: 12px;
+    height: 100%;
+
+    ${LoggerInput} {
+      height: 100%;
+      overflow-y: scroll;
+    }
+
+    .actions {
+      display: flex;
+      flex-direction: row;
+      gap: 8px;
+      justify-content: flex-end;
+    }
+  }
+
+  .success {
+    display: flex;
+    gap: 16px;
+    justify-content: space-between;
+    align-items: center;
+    padding: 8px 16px 8px 16px;
+    border-radius: 8px;
+    border: 1px solid ${({ theme }) => theme.colors.successBase};
+    background-color: ${({ theme }) => theme.colors.successSoft};
+    min-height: 56px;
+    box-sizing: border-box;
+
+    .icon-message {
+      display: flex;
+      gap: 16px;
+      color: ${({ theme }) => theme.colors.successBase};
+      align-items: center;
+
+      .message {
+        ${({ theme }) => theme.fonts.normalRegular}
+        color: ${({ theme }) => theme.colors.text100};
+      }
+    }
+
+    .show-folder {
+      ${({ theme }) => theme.fonts.normalMedium}
+      color: ${({ theme }) => theme.colors.successBase};
+
+      :hover {
+        cursor: pointer;
+      }
+    }
+  }
+
+  .error {
+    ${({ theme }) => theme.fonts.normalRegular}
+    color: ${({ theme }) => theme.colors.dangerBase};
+  }
+`;
+
+type ProgressBarCompilationContainerProps = {
+  isError: boolean;
+};
+
+export const ProgressBarCompilationContainer = styled(BlockProgressBar)<ProgressBarCompilationContainerProps>`
+  justify-content: initial;
+  gap: 12px;
+  width: 100%;
+
+  .progress {
+    width: 100%;
+    height: 12px;
+  }
+
+  .progress::-webkit-progress-value {
+    background: ${({ theme, isError }) => (isError ? theme.colors.dangerBase : theme.colors.successBase)};
+  }
+
+  .progress-message {
+    ${({ theme, isError }) => (isError ? theme.fonts.normalMedium : theme.fonts.normalRegular)}
+    color: ${({ theme, isError }) => (isError ? theme.colors.dangerBase : theme.colors.text100)};
+  }
+
+  .platform {
+    ${({ theme }) => theme.fonts.normalMedium}
+    color: ${({ theme }) => theme.colors.text100};
   }
 `;
